requirements.txt
Authomatic==0.1.0.post1
Fabric==1.12.0
Flask==0.9
Jinja2==2.7.3
MarkupSafe==0.23
PyYAML==3.11
Werkzeug==0.8.3
bcdoc==0.12.0
boto==2.34.0
botocore==1.4.56
clan==0.1.3
-e git+https://github.com/nprapps/copydoc.git#egg=copydoc
-e git+https://github.com/nprapps/shortcodes.git@python2#egg=shortcodes
copytext==0.1.9
docutils==0.11
gunicorn==19.1.1
httplib2==0.9
jmespath==0.2.1
nose==1.2.1
odict==1.5.1
openpyxl==2.2.0
Pillow==3.4.2
ply==3.4
pyasn1==0.1.7
pycrypto==2.6
python-dateutil==2.2
requests==2.5.0
slimit==0.7.4
smartypants==1.8.6
ssh==1.7.14
termcolor==1.1.0
wheel==0.23.0
wsgiref==0.1.2
